New path '/foroApi/authCookies' added. Its handler set the username and user_session of the  on the cookie storage of the client.
New path '/foroApi/logout' added.
Its handler set the cookies, username and user_session, of the client to '', or in other words it log out the client.

The handler of the context property has changued.
Now it is responsible to read the cookies of all of each client requests and see if there is  the user_session (jwt token) and try to decipher it. After trying to decipher it, if there is a value then it return the same user_session and username cookies. if there is not a value then it logout the client.