After all the tests made on the forTestDeployOnHeroku branch, i know how to set cookies on a response and thanks to that create a new way to validate a user session and keep it alive. The session starts with a requests to '/foroApi/signin', and if the username and password are valid then the server send a response 200 with the authentication cookies of the user. With the authentication cookies on the client cookies storage, all the request, sent from the client, are going to have that cookies.
Thanks to the context property, the server checks if the authentication cookie 'user_session' belongs to a user. If it belongs to a user the server sends a response with an status code of 203 to tell the client the user is authenticated, if it doesnt belongs to a user then the server delete the cookies of the client and send response with an status code of 401 to tell the client the user is not authenticated.